Add portgroups to support LAG interfaces - RPC
Ironic should be able to provide the requisite connectivity
information to the Neutron ML2 plugin to allow drivers to
provision the top-of-rack switch for the bare metal server.
The addition of portgroups in Ironic allows the concept of
link aggregation to be handled in Ironic in order to provide
support for cases where multiple interfaces on the bare metal
server connect to switch ports of a single LAG.

This commit includes changes to:
- the RPC API (to include portgroup methods)
- the RPC tests

Spec available at:
http://specs.openstack.org/openstack/ironic-specs/specs/liberty/
ironic-ml2-integration.html
Note: this is an initial draft only and is still work in progress.

Implements: blueprint ironic-ml2-integration
DocImpact
Co-Authored-By: Jenny Moorehead (jenny.moorehead@sap.com)
Co-Authored-By: Will Stevenson (will.stevenson@sap.com)

Change-Id: Iac6355e615e57e6ed002facd6795c9ff06188998